식단 및 레시피 (MenuRecipe) 파일 부분

+++ schoolMeal/src/main/java/com/example/schoolMeal/domain/repository/mealResource/MenuRecipeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.schoolMeal.domain.entity.mealResource.MenuRecipe;
 
 @Repository
-public interface MenuRecipeRepository extends JpaRepository<MenuRecipe, Integer> {
+public interface MenuRecipeRepository extends JpaRepository<MenuRecipe, Long> {
 
 }
